requirements.txt
fastapi
uvicorn
python-dotenv
httpx
supabase
pydantic-ai
pydantic
requests